extract export_import classes
