+++ lib/components/dialoges/enter_timeformat.dart
@@ -1,4 +1,4 @@
-import 'package:blood_pressure_app/screens/subsettings/export_field_format_documentation.dart';
+import 'package:blood_pressure_app/screens/subsettings/export_import/export_field_format_documentation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
