+++ lib/l10n/app_en.arb
@@ -221,7 +221,7 @@
       }
     }
   },
-  "exportWarnConfigNotImportable": "Hey! Just a friendly heads up: the current export configuration won''t be importable. To fix it, make sure you set the export type as CSV, enable the headline, and include one of the time formats available.",
+  "exportWarnConfigNotImportable": "Hey! Just a friendly heads up: the current export configuration won''t be importable. To fix it, make sure you set the export type as CSV and include one of the time formats available.",
   "@exportWarnConfigNotImportable": {},
   "exportWarnNotEveryFieldExported": "Beware that you are not exporting all fields: {fields} {count, plural, one{is} other{are}} missing.",
   "@exportWarnNotEveryFieldExported": {
@@ -474,5 +474,7 @@
   "exportFieldsPreset": "Export fields preset",
   "@exportFieldsPreset": {},
   "remove": "Remove",
-  "@remove": {}
+  "@remove": {},
+  "manageExportColumns": "Manage export columns",
+  "@manageExportColumns": {}
 }
